Update Pterodactyl.php
I delved into the code process, and realized that this was done absolutely contradictoryly all the rules and norms of the code. This is the simplest and work option that will completely remove the error when creating a user, regardless of the user registration language. In FIRST and LAST NAME! The mail all will be true in the name of the login and it will be possible not to be exhausted above the code!