Docker-Compose & README Changes (#792)
* Update docker-compose.beta.yml

Version is obsolete, we can see this here: https://docs.docker.com/reference/compose-file/version-and-name/

Changed the mariadb image to lts (long-time-support) because we are using an old version of this image.

* Update docker-compose.example.yml

Version is obsolete, we can see this here: https://docs.docker.com/reference/compose-file/version-and-name/

Changed the mariadb image to lts (long-time-support) because we are using an old version of this image.

* Update README.md

Not everyone checks the website directly, so I’ve added the extensions to the README. This way, users can see all relevant information directly within the GitHub repository.

* Update README.md

Added some more requirements.