Update DirectAdmin.php
I fixed a security issue in a direct administrator extension for SSO login. I modified the API call for that API call alone so that it actually generates the key for the user itself. He still logged in with the administrator account, but I had to make a new API request funtion specifically for SSO login to use the user login.